Generalize patch-eslint6.js to work with ESLint 7
